cell configuration logic 별도 파일로 분리
